Optimize visitor count APIs to avoid fetching all records

+++ src/main/java/com/coderprabhu/api/controllers/VisitorController.java
@@ -31,8 +31,8 @@ public class VisitorController {
 
     @GetMapping(value="/count")
     @CrossOrigin(origins = "*")
-    public int count() {
-        Integer totalVisits = visitorService.getTotalVisits();
+    public long count() {
+        long totalVisits = visitorService.getTotalVisits();
         log.info("Total visitors so far: " + totalVisits);
         return totalVisits;
     }
